requirements.txt
python-dotenv
ccxt
aiohttp
requests
python-telegram-bot
